+++ crates/router/src/connector/utils.rs
@@ -50,6 +50,7 @@ pub trait RouterData {
     fn get_billing_country(&self) -> Result<api_models::enums::CountryCode, Error>;
     fn get_billing_phone(&self) -> Result<&api::PhoneDetails, Error>;
     fn get_description(&self) -> Result<String, Error>;
+    fn get_return_url(&self) -> Result<String, Error>;
     fn get_billing_address(&self) -> Result<&api::AddressDetails, Error>;
     fn get_shipping_address(&self) -> Result<&api::AddressDetails, Error>;
     fn get_connector_meta(&self) -> Result<pii::SecretSerdeValue, Error>;
@@ -88,6 +89,11 @@ impl<Flow, Request, Response> RouterData for types::RouterData<Flow, Request, Re
         self.description
             .clone()
             .ok_or_else(missing_field_err("description"))
+    }
+    fn get_return_url(&self) -> Result<String, Error> {
+        self.return_url
+            .clone()
+            .ok_or_else(missing_field_err("return_url"))
     }
     fn get_billing_address(&self) -> Result<&api::AddressDetails, Error> {
         self.address
@@ -139,6 +145,8 @@ pub trait PaymentsAuthorizeRequestData {
     fn get_browser_info(&self) -> Result<types::BrowserInformation, Error>;
     fn get_card(&self) -> Result<api::Card, Error>;
     fn get_return_url(&self) -> Result<String, Error>;
+    fn is_mandate_payment(&self) -> bool;
+    fn get_webhook_url(&self) -> Result<String, Error>;
 }
 
 impl PaymentsAuthorizeRequestData for types::PaymentsAuthorizeData {
@@ -163,6 +171,19 @@ impl PaymentsAuthorizeRequestData for types::PaymentsAuthorizeData {
         self.router_return_url
             .clone()
             .ok_or_else(missing_field_err("return_url"))
+    }
+    fn is_mandate_payment(&self) -> bool {
+        self.setup_mandate_details.is_some()
+            || self
+                .mandate_id
+                .as_ref()
+                .and_then(|mandate_ids| mandate_ids.connector_mandate_id.as_ref())
+                .is_some()
+    }
+    fn get_webhook_url(&self) -> Result<String, Error> {
+        self.router_return_url
+            .clone()
+            .ok_or_else(missing_field_err("webhook_url"))
     }
 }
 
@@ -247,7 +268,11 @@ pub enum CardIssuer {
 pub trait CardData {
     fn get_card_expiry_year_2_digit(&self) -> Secret<String>;
     fn get_card_issuer(&self) -> Result<CardIssuer, Error>;
-    fn get_card_expiry_month_year_2_digit_with_delimiter(&self, delimiter: String) -> String;
+    fn get_card_expiry_month_year_2_digit_with_delimiter(
+        &self,
+        delimiter: String,
+    ) -> Secret<String>;
+    fn get_expiry_date_as_yyyymm(&self, delimiter: &str) -> Secret<String>;
 }
 
 impl CardData for api::Card {
@@ -263,14 +288,29 @@ impl CardData for api::Card {
             .map(|card| card.split_whitespace().collect());
         get_card_issuer(card.peek().clone().as_str())
     }
-    fn get_card_expiry_month_year_2_digit_with_delimiter(&self, delimiter: String) -> String {
+    fn get_card_expiry_month_year_2_digit_with_delimiter(
+        &self,
+        delimiter: String,
+    ) -> Secret<String> {
         let year = self.get_card_expiry_year_2_digit();
-        format!(
+        Secret::new(format!(
             "{}{}{}",
             self.card_exp_month.peek().clone(),
             delimiter,
             year.peek()
-        )
+        ))
+    }
+    fn get_expiry_date_as_yyyymm(&self, delimiter: &str) -> Secret<String> {
+        let mut x = self.card_exp_year.peek().clone();
+        if x.len() == 2 {
+            x = format!("20{}", x);
+        }
+        Secret::new(format!(
+            "{}{}{}",
+            x,
+            delimiter,
+            self.card_exp_month.peek().clone()
+        ))
     }
 }
 
@@ -453,15 +493,6 @@ impl common_utils::errors::ErrorSwitch<errors::ConnectorError> for errors::Parsi
     fn switch(&self) -> errors::ConnectorError {
         errors::ConnectorError::ParsingFailed
     }
-}
-
-pub fn to_string<T>(data: &T) -> Result<String, Error>
-where
-    T: serde::Serialize,
-{
-    serde_json::to_string(data)
-        .into_report()
-        .change_context(errors::ConnectorError::ResponseHandlingFailed)
 }
 
 pub fn base64_decode(data: String) -> Result<Vec<u8>, Error> {