+++ crates/router/src/core/payments/operations/payment_capture.rs
@@ -160,6 +160,7 @@ impl<F: Send + Clone> GetTracker<F, payments::PaymentData<F>, api::PaymentsCaptu
                 sessions_token: vec![],
                 card_cvc: None,
                 creds_identifier,
+                pm_token: None,
             },
             None,
         ))