+++ crates/router/src/types/storage/payment_attempt.rs
@@ -1,41 +1,11 @@
-use error_stack::ResultExt;
 pub use storage_models::payment_attempt::{
     PaymentAttempt, PaymentAttemptNew, PaymentAttemptUpdate, PaymentAttemptUpdateInternal,
-};
-
-use crate::{
-    core::errors::{self, CustomResult},
-    utils::ValueExt,
 };
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct RoutingData {
     pub routed_through: Option<String>,
     pub algorithm: Option<api_models::admin::RoutingAlgorithm>,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct RoutedThroughData {
-    pub routed_through: Option<String>,
-}
-
-pub trait PaymentAttemptExt {
-    fn get_routed_through_connector(&self) -> CustomResult<Option<String>, errors::ParsingError>;
-}
-
-impl PaymentAttemptExt for PaymentAttempt {
-    fn get_routed_through_connector(&self) -> CustomResult<Option<String>, errors::ParsingError> {
-        if let Some(ref val) = self.connector {
-            let data: RoutedThroughData = val
-                .clone()
-                .parse_value("RoutedThroughData")
-                .attach_printable("Failed to read routed_through connector from payment attempt")?;
-
-            Ok(data.routed_through)
-        } else {
-            Ok(None)
-        }
-    }
 }
 
 #[cfg(feature = "kv_store")]
@@ -67,9 +37,7 @@ mod tests {
         let connector = types::Connector::Dummy.to_string();
         let payment_attempt = PaymentAttemptNew {
             payment_id: payment_id.clone(),
-            connector: Some(serde_json::json!({
-                "routed_through": connector,
-            })),
+            connector: Some(connector),
             created_at: current_time.into(),
             modified_at: current_time.into(),
             ..PaymentAttemptNew::default()
@@ -102,9 +70,7 @@ mod tests {
         let payment_attempt = PaymentAttemptNew {
             payment_id: payment_id.clone(),
             merchant_id: merchant_id.clone(),
-            connector: Some(serde_json::json!({
-                "routed_through": connector,
-            })),
+            connector: Some(connector),
             created_at: current_time.into(),
             modified_at: current_time.into(),
             attempt_id: attempt_id.clone(),
@@ -146,9 +112,7 @@ mod tests {
         let payment_attempt = PaymentAttemptNew {
             payment_id: uuid.clone(),
             merchant_id: "1".to_string(),
-            connector: Some(serde_json::json!({
-                "routed_through": connector,
-            })),
+            connector: Some(connector),
             created_at: current_time.into(),
             modified_at: current_time.into(),
             // Adding a mandate_id