+++ crates/storage_models/src/schema.rs
@@ -222,6 +222,7 @@ diesel::table! {
         locker_id -> Nullable<Varchar>,
         metadata -> Nullable<Jsonb>,
         routing_algorithm -> Nullable<Json>,
+        primary_business_details -> Json,
         api_key -> Nullable<Varchar>,
     }
 }
@@ -241,6 +242,10 @@ diesel::table! {
         payment_methods_enabled -> Nullable<Array<Nullable<Json>>>,
         connector_type -> ConnectorType,
         metadata -> Nullable<Jsonb>,
+        connector_label -> Varchar,
+        business_country -> CountryCode,
+        business_label -> Varchar,
+        business_sub_label -> Nullable<Varchar>,
     }
 }
 
@@ -257,7 +262,7 @@ diesel::table! {
         amount -> Int8,
         currency -> Nullable<Currency>,
         save_to_locker -> Nullable<Bool>,
-        connector -> Nullable<Jsonb>,
+        connector -> Nullable<Varchar>,
         error_message -> Nullable<Text>,
         offer_amount -> Nullable<Int8>,
         surcharge_amount -> Nullable<Int8>,
@@ -282,6 +287,8 @@ diesel::table! {
         payment_experience -> Nullable<Varchar>,
         payment_method_type -> Nullable<Varchar>,
         payment_method_data -> Nullable<Jsonb>,
+        business_sub_label -> Nullable<Varchar>,
+        straight_through_algorithm -> Nullable<Jsonb>,
     }
 }
 
@@ -313,6 +320,8 @@ diesel::table! {
         off_session -> Nullable<Bool>,
         client_secret -> Nullable<Varchar>,
         active_attempt_id -> Varchar,
+        business_country -> CountryCode,
+        business_label -> Varchar,
     }
 }