+++ CHANGELOG.md
@@ -1,3 +1,49 @@
+# 0.5.5 (2023-04-10)
+
+## New Features
+
+* **api_models:**  derive `strum::Display` for `RefundStatus` (#846) (4524d4f5)
+*  allow (de)serializing countries to/from alpha-2, alpha-3 and numeric country codes (#836) (899767cf)
+* **connector:**  add authorize, capture, void, psync, refund, rsync for PayPal connector (#747) (36049c13)
+
+## Bug Fixes
+
+*  Add locker sign keyid in env (#844) (70dff140)
+
+## Other Changes
+
+* **common_utils:**  put the async ext trait behind a feature (#835) (de29eb68)
+*  update ci workflows for common_enums crate (#843) (45111337)
+
+## Refactors
+
+* **scheduler:**  remove scheduler options & adding graceful shutdown to producer (#840) (11df8436)
+* **router:**  refactor amount in PaymentsCaptureData from Option<i64> to i64 (#821) (b8bcba4e)
+
+# 0.5.4 (2023-04-04)
+
+## New Features
+
+* **request:**  add `RequestBuilder` method to attach default request headers (#826) (6f61f830)
+* **middleware:**  add middleware to attach default response headers (#824) (6d7b11a0)
+* **core:**  added multiple payment_attempt support for payment_intent (#439) (35d3e277)
+* **router:**  added incoming dispute webhooks flow (#769) (a733eafb)
+
+## Bug Fixes
+
+* **cards_info:**  add extra columns to cards_info struct (#813) (442bed0f)
+* **connector:**  [Mollie] remove unsupported implementation of Void flow from mollie connector (#808) (eee8304b)
+
+## Other Changes
+
+* **common_utils:**  put the signals module behind a feature flag (#814) (fb4ec431)
+* **core:**  replace string with enum for country (#735) (e18bfb2a)
+* **api_models:**  put the errors module behind a feature flag (#815) (f14f87a1)
+* **storage_models:**  delete client secret when status is succeeded, failed, cancelled (#724) (a05059b7)
+## Refactors
+
+* **drainer, router:**  KMS decrypt database password when `kms` feature is enabled (#733) (9d6e4ee3)
+
 # 0.5.3 (2023-03-29)
 
 ## Documentation Changes